# docs/swagger.yaml
definitions:
  models.ClinicalNote:
    properties:
      consultationVisit:
        $ref: '#/definitions/models.ConsultationVisit'
      id_consulta:
        type: integer
      notas_progreso:
        type: string
      observaciones_recomendaciones:
        type: string
    type: object
...
  models.ConsultationVisit:
    properties:
      fecha_hora_visita:
        type: string
      hospitalEmployee:
        $ref: '#/definitions/models.HospitalEmployee'
      id_paciente:
        type: integer
      id_personal_hospital:
        type: integer
      notas_medicas:
        type: string
      patient:
        $ref: '#/definitions/models.Patient'
      razon_visita:
        type: string
      resultados_examenes:
        type: string
    type: object
  models.DemographicData:
    properties:
      estado_civil:
        type: string
      informacion_etnica_racial:
        type: string
      informacion_laboral:
        type: string
      nacionalidad:
        type: string
    type: object
...
  models.HospitalEmployee:
    properties:
      direccion:
        type: string
      fecha_nacimiento:
        type: string
      genero:
        type: string
      id_identificacion:
        type: integer
      id_rol:
        description: Llave foránea que referencia la tabla roles
        type: integer
      identification:
        $ref: '#/definitions/models.Identification'
      nombre_completo:
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.Role'
        description: Carga la información del rol relacionado
      telefono:
        type: string
    type: object
  models.Identification:
    properties:
      numero_documento:
        type: string
      tipo_documento:
        type: string
    type: object
...
  models.Patient:
    properties:
      demographicData:
        $ref: '#/definitions/models.DemographicData'
      direccion:
        type: string
      fecha_nacimiento:
        type: string
      genero:
        type: string
      id_datos_demograficos:
        type: integer
      id_identificacion:
        type: integer
      id_rol:
        description: Llave foránea que referencia la tabla roles
        type: integer
      id_usuarios:
        type: integer
      identification:
        $ref: '#/definitions/models.Identification'
      nombre_completo:
        type: string
      numero_seguro_social:
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.Role'
        description: Carga la información del rol relacionado
      telefono:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.Role:
    properties:
      descripcion:
        type: string
      nombre_rol:
        type: string
    type: object
...
  models.User:
    properties:
      contraseña:
        type: string
      id_rol:
        description: Llave foránea que referencia la tabla roles
        type: integer
      nombre_usuario:
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.Role'
        description: Carga la información del rol relacionado
    type: object
  request.CreateClinicalNoteDTO:
    properties:
      id_consulta:
        type: integer
      notas_progreso:
        type: string
      observaciones_recomendaciones:
        type: string
    required:
    - id_consulta
    - notas_progreso
    - observaciones_recomendaciones
    type: object
...
  request.CreateConsultationVisitDTO:
    properties:
      fecha_hora_visita:
        type: string
      id_paciente:
        type: integer
      id_personal_hospital:
        type: integer
      notas_medicas:
        type: string
      razon_visita:
        type: string
      resultados_examenes:
        type: string
    required:
    - fecha_hora_visita
    - id_paciente
    - id_personal_hospital
    - notas_medicas
    - razon_visita
    - resultados_examenes
    type: object
...
info:
  contact: {}
paths:
  /clinical_notes:
    get:
      description: Retrieve a list of all clinical_notes in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ClinicalNote'
            type: array
      summary: List all clinical_notes
      tags:
      - clinical_notes
    post:
      consumes:
      - application/json
      description: Adds a new clinical_note to the system
      parameters:
      - description: ClinicalNote data
        in: body
        name: clinical_note
        required: true
        schema:
          $ref: '#/definitions/request.CreateClinicalNoteDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ClinicalNote'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new clinical_note
      tags:
      - clinical_notes
  /clinical_notes/{id}:
    delete:
      description: Remove a clinical_note by its ID
      parameters:
      - description: ClinicalNote ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a clinical_note
      tags:
      - clinical_notes
    get:
      description: Retrieve a single clinical_note by its ID
      parameters:
      - description: ClinicalNote ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ClinicalNote'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a clinical_note by ID
      tags:
      - clinical_notes
    put:
      consumes:
      - application/json
      description: Update the information of an existing clinical_note
      parameters:
      - description: ClinicalNote ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated clinical_note data
        in: body
        name: clinical_note
        required: true
        schema:
          $ref: '#/definitions/request.CreateClinicalNoteDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ClinicalNote'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a clinical_note
      tags:
      - clinical_notes
...
  /consultation_visits:
    get:
      description: Retrieve a list of all consultation_visits in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ConsultationVisit'
            type: array
      summary: List all consultation_visits
      tags:
      - consultation_visits
    post:
      consumes:
      - application/json
      description: Adds a new consultation_visit to the system
      parameters:
      - description: ConsultationVisit data
        in: body
        name: consultation_visit
        required: true
        schema:
          $ref: '#/definitions/request.CreateConsultationVisitDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ConsultationVisit'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new consultation_visit
      tags:
      - consultation_visits
  /consultation_visits/{id}:
    delete:
      description: Remove a consultation_visit by its ID
      parameters:
      - description: ConsultationVisit ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a consultation_visit
      tags:
      - consultation_visits
    get:
      description: Retrieve a single consultation_visit by its ID
      parameters:
      - description: ConsultationVisit ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ConsultationVisit'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a consultation_visit by ID
      tags:
      - consultation_visits
    put:
      consumes:
      - application/json
      description: Update the information of an existing consultation_visit
      parameters:
      - description: ConsultationVisit ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated consultation_visit data
        in: body
        name: consultation_visit
        required: true
        schema:
          $ref: '#/definitions/request.CreateConsultationVisitDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ConsultationVisit'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a consultation_visit
      tags:
      - consultation_visits
...
swagger: "2.0"